--- resources/assets/js/components/movies/filters/Filter.js
@@ -1,21 +1,0 @@
-import React, { PureComponent } from 'react';
-import styled from 'styled-components';
-import { rem } from 'polished';
-import { colors, styles } from '../../shared';
-
-import FilterItem from './item';
-
-const FilterBar = styles.Container.extend``;
-
-export default class Filter extends PureComponent {
-  render() {
-    return(
-      <FilterBar data-test="filter-bar">
-        <FilterItem
-          criterion="actors"
-          data-test="filter-item"
-        />
-      </FilterBar>
-    );
-  }
-}
