+++ resources/assets/js/components/movies/filters/__tests__/Filter.test.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Filter from '../Filter';
+import Filter from '../';
 
 jest.mock('../item', () => () => <div />);
